Restructure code organization
